requirements.txt
msal
pyautogui
python-dotenv
selenium
webdriver_manager